build: revert FileAccess/FileInfo/serialize split back to master
The serializer stays in the runtime lib (Program::serialize is a member of
the runtime-exported Program and can't be DAS_CC_API-exported on MSVC), so
splitting FileAccess/FileInfo and the serializer across the runtime/compiler
boundary served no purpose. Revert that whole effort to master:

- FileAccess/FileInfo/TextFileInfo/ModuleFileAccess::serialize and the
  FileInfo */FileAccessPtr serialize operators return to their master form in
  module_builtin_ast_serialize.cpp (virtual serialize, no serializeKind tag).
- FileAccess / AstSerializer daslang bindings move back rtti -> rtti
  (module_builtin_rtti.cpp/.h, daslib/rtti.das); the typedef FileAccessPtr
  returns to rtti.das.
- FileAccess base impl returns to src/simulate/debug_info.cpp;
  src/simulate/file_access.cpp is deleted.
- ast_serializer.h rtti_* serializer decls drop the DAS_API requalification.

All reverted files are byte-identical to master. Kept: the unrelated
setRefSp visitor-dispatch rewrite and the ast_derive_alias/ast_gc_report/
ast_inscope_pod/ast_escape_analysis/ast_validate moves to the compiler lib.

Verified: build clean (static + Dyn shared libs); serialize/deserialize
round-trip over ./tests 8899/8899.
