+++ include/daScript/ast/ast_serializer.h
@@ -334,22 +334,28 @@ namespace das {
         unique_ptr<AstSerializer> serializer;
     };
 
-    DAS_API AstSerializerState * rtti_create_ast_serializer ();
+    // Create a writing serializer.
+    AstSerializerState * rtti_create_ast_serializer ();
 
-    DAS_API AstSerializerState * rtti_create_ast_deserializer ( const TArray<uint8_t> & data );
+    // Create a reading serializer from a blob.
+    AstSerializerState * rtti_create_ast_deserializer ( const TArray<uint8_t> & data );
 
-    DAS_API void rtti_delete_ast_serializer ( AstSerializerState * state );
+    // Delete a serializer state.
+    void rtti_delete_ast_serializer ( AstSerializerState * state );
 
-    DAS_API bool rtti_ast_serializer_serialize_program (
+    // Serialize one program (writing mode). Returns false on failure.
+    bool rtti_ast_serializer_serialize_program (
             AstSerializerState * state,
             const smart_ptr<Program> & program );
 
-    DAS_API void rtti_ast_serializer_deserialize_program (
+    // Deserialize one program (reading mode).
+    void rtti_ast_serializer_deserialize_program (
             AstSerializerState * state,
             const TBlock<void,bool,smart_ptr<Program>,const string> & block,
             Context * context, LineInfoArg * at );
 
-    DAS_API void rtti_ast_serializer_get_data (
+    // Get serialized data from a writing serializer.
+    void rtti_ast_serializer_get_data (
             AstSerializerState * state,
             const TBlock<void,TTemporary<TArray<uint8_t> const>> & block,
             Context * context, LineInfoArg * at );
