+++ src/builtin/module_builtin_ast_serialize.cpp
@@ -369,10 +369,6 @@ namespace das {
         return *this;
     }
 
-    // explicit instantiation used by FileAccess::serialize in module_file_access.cpp
-    template AstSerializer & AstSerializer::operator << (
-        das_hash_map<string, unique_ptr<FileInfo>, daslang_hash<string,void>, das::equal_to<string>> & );
-
     // Guarded: see ast_serializer.h note. Under DAS_CUSTOM_HASH=0 these would
     // duplicate the das_hash_map definitions above (identical aliases).
 #if DAS_CUSTOM_HASH
@@ -796,18 +792,6 @@ namespace das {
         return *this;
     }
 
-    // FileInfo serialize used to be a virtual on FileInfo / TextFileInfo. Dropped
-    // so FileInfo's vtable doesn't reference a compiler-only symbol (this TU is
-    // now compiler-only; FileInfo lives in runtime). Dispatch by class tag here.
-    static void serializeFileInfoBody ( AstSerializer & ser, FileInfo * info, uint8_t tag ) {
-        ser << info->name << info->tabSize;
-        if ( tag == 1 ) {
-            auto tf = static_cast<TextFileInfo*>(info);
-            ser.serializeAdaptiveSize32(tf->sourceLength);
-        }
-        if ( !ser.writing ) ser.deleteUponFinish.push_back(info);
-    }
-
     AstSerializer & AstSerializer::operator << ( FileInfo * & info ) {
         dtag(HASH_TAG("FileInfo *"));
         bool is_null = info == nullptr;
@@ -821,9 +805,7 @@ namespace das {
                 uint64_t curOffset = buffer->writingSize() + sizeof(curOffset);
                 *this << curOffset;
                 writingFileInfoMap[info] = curOffset;
-                uint8_t tag = info->serializeKind;
-                *this << tag;
-                serializeFileInfoBody(*this, info, tag);
+                info->serialize(*this);
             } else {
                 *this << writingFileInfoMap[info];
             }
@@ -838,7 +820,7 @@ namespace das {
                     case 1: info = new TextFileInfo; break;
                     default: SERIALIZER_VERIFYF(false, "Unreachable");
                 }
-                serializeFileInfoBody(*this, info, tag);
+                info->serialize(*this);
                 readingFileInfoMap[curOffset] = info;
                 if ( curOffset != savedOffset )
                     readOffset = savedOffset;
@@ -887,12 +869,20 @@ namespace das {
         return *this;
     }
 
-    // FileAccess serialize used to be a virtual on FileAccess / ModuleFileAccess.
-    // Dropped so FileAccess's vtable doesn't reference a compiler-only symbol
-    // (this TU is compiler-only; FileAccess lives in runtime). The class tag is
-    // ptr->serializeKind, written/read by the operator below.
-    static void serializeFileAccessBody ( AstSerializer & ser, FileAccess * fa ) {
-        ser << fa->files;
+    void FileAccess::serialize ( AstSerializer & ser ) {
+        if ( ser.writing ) {
+            uint8_t tag = 0;
+            ser << tag;
+        }
+        ser << files;
+    }
+
+    void ModuleFileAccess::serialize ( AstSerializer & ser ) {
+        if ( ser.writing ) {
+            uint8_t tag = 1;
+            ser << tag;
+        }
+        ser << files;
     }
 
     AstSerializer & AstSerializer::operator << ( FileAccessPtr & ptr ) {
@@ -908,8 +898,7 @@ namespace das {
             *this << p;
             if ( fileAccessMap[p] == nullptr ) {
                 fileAccessMap[p] = ptr.get();
-                *this << ptr->serializeKind;
-                serializeFileAccessBody(*this, ptr.get());
+                ptr->serialize(*this);
             }
         } else {
             SerializeNodeId p; *this << p;
@@ -921,7 +910,7 @@ namespace das {
                     default: SERIALIZER_VERIFYF(false, "Unreachable");
                 }
                 fileAccessMap[p] = ptr.get();
-                serializeFileAccessBody(*this, ptr.get());
+                ptr->serialize(*this);
             } else {
                 ptr.orphan();
                 FileAccessPtr t = fileAccessMap[p];
@@ -1987,6 +1976,37 @@ namespace das {
 
     void MakeStruct::serialize( AstSerializer & ser ) {
         ser << static_cast <vector<MakeFieldDeclPtr> &> ( *this );
+    }
+
+    void FileInfo::serialize ( AstSerializer & ser ) {
+        uint8_t tag = 0;
+        if ( ser.writing ) {
+            ser << tag;
+        }
+        ser << name << tabSize;
+        if ( !ser.writing ) {
+            ser.deleteUponFinish.push_back(this);
+        }
+        // Note: we do not serialize profileData
+    }
+
+    void TextFileInfo::serialize ( AstSerializer & ser ) {
+        uint8_t tag = 1; // Signify the text file info
+        if ( ser.writing ) {
+            ser << tag;
+        }
+        ser << name << tabSize;
+        ser.serializeAdaptiveSize32(sourceLength);
+        if ( !ser.writing ) {
+            ser.deleteUponFinish.push_back(this);
+        }
+        // ser << owner;
+        // if ( ser.writing ) {
+        //     ser.write((const void *) source, sourceLength);
+        // } else {
+        //     source = (char *) das_aligned_alloc16(sourceLength + 1);
+        //     ser.read((void *) source, sourceLength);
+        // }
     }
 
     AstSerializer & AstSerializer::operator << ( CallMacro * & ptr ) {
